# Hustle-Safe/Hustle-Safe/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HustleSafe API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: List policies
      parameters:
        - name: worker_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of policies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Create a new policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      summary: Get a policy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List all claims
      parameters:
        - name: worker_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: zone_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimListResponse"
  /claims/{id}:
    get:
      operationId: getClaim
      tags: [claims]
      summary: Get a claim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Claim details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
    patch:
      operationId: updateClaim
      tags: [claims]
      summary: Update claim status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClaimRequest"
      responses:
        "200":
          description: Updated claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
...
  /fraud/evaluate:
    post:
      operationId: evaluateFraudClaim
      tags: [fraud]
      summary: Evaluate fraud for a worker claim context using full adaptive pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FraudEvaluationRequest"
      responses:
        "200":
          description: Final Adaptive Fraud Evaluation Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FraudEvaluationResult"
  /fraud/audit/{claim_id}:
    get:
      operationId: getFraudAuditLog
      tags: [fraud]
      summary: Fetch the full audit log and XGBoost breakdown for a claim
      parameters:
        - name: claim_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Audit log response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudAuditLog"
  /fraud/signals/{worker_id}:
    get:
      operationId: getWorkerSignalHistory
      tags: [fraud]
      summary: Fetch recent signals and fraud evaluations mapped to a worker ID
      parameters:
        - name: worker_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of signal tracking history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FraudAuditLog"
components:
  schemas:
...
    Policy:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        tier:
          type: string
          enum: [basic, standard, pro]
        weekly_premium:
          type: number
        coverage_cap:
          type: number
        status:
          type: string
          enum: [active, inactive, suspended]
        activated_at:
          type: string
        expires_at:
          type: string
          nullable: true
        zone_id:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - worker_id
        - tier
        - weekly_premium
        - coverage_cap
        - status
        - activated_at
        - created_at
    PolicyListResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
        total:
          type: integer
      required:
        - policies
        - total
    CreatePolicyRequest:
      type: object
      properties:
        worker_id:
          type: string
        tier:
          type: string
          enum: [basic, standard, pro]
        zone_id:
          type: string
      required:
        - worker_id
        - tier
    Claim:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        policy_id:
          type: string
          nullable: true
        zone_id:
          type: string
          nullable: true
        disruption_type:
          type: string
          nullable: true
        disruption_start:
          type: string
          nullable: true
        disruption_end:
          type: string
          nullable: true
        hours_affected:
          type: number
          nullable: true
        hourly_rate:
          type: number
        payout_amount:
          type: number
          nullable: true
        fraud_score:
          type: number
        status:
          type: string
          enum: [pending, auto_approved, soft_hold, insurer_review, auto_rejected, paid]
        fraud_signals:
          type: object
          additionalProperties: true
        resolution_notes:
          type: string
          nullable: true
        created_at:
          type: string
        paid_at:
          type: string
          nullable: true
        worker_name:
          type: string
          nullable: true
        zone_name:
          type: string
          nullable: true
      required:
        - id
        - worker_id
        - hourly_rate
        - fraud_score
        - status
        - fraud_signals
        - created_at
    ClaimListResponse:
      type: object
      properties:
        claims:
          type: array
          items:
            $ref: "#/components/schemas/Claim"
        total:
          type: integer
      required:
        - claims
        - total
    UpdateClaimRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, auto_approved, soft_hold, insurer_review, auto_rejected, paid]
        resolution_notes:
          type: string
...
    FraudEvaluationRequest:
      type: object
      required:
        - worker_id
        - zone_id
        - user_lat
        - user_lng
        - device_id
      properties:
        worker_id:
          type: string
          format: uuid
          description: Unique identifier of the worker.
        zone_id:
          type: string
          description: Identifier of the zone the worker claims to be disrupted in.
        user_lat:
          type: number
          format: float
          description: Worker's current latitude.
        user_lng:
          type: number
          format: float
          description: Worker's current longitude.
        device_id:
          type: string
          description: Fingerprint ID of the user's mobile device for fraud checks.
...
    FraudEvaluationResult:
      type: object
      required:
        - final_score
        - resolution_tier
        - individual_signal_scores
        - signal_status
        - critical_flags_fired
        - critical_flag_signals
        - worker_trust_score
        - adaptive_threshold_used
        - primary_elevated_signal
        - pass2_triggered
        - pass2_trigger_reason
        - adversarial_elevated_count
        - data_quality
      properties:
        final_score:
          type: number
          format: float
        resolution_tier:
          type: string
        individual_signal_scores:
          type: object
          additionalProperties:
            type: number
            nullable: true
        signal_status:
          type: object
          additionalProperties:
            type: string
            enum: [computed, missing_data, error]
        critical_flags_fired:
          type: boolean
        critical_flag_signals:
          type: array
          items:
            type: string
        worker_trust_score:
          type: number
          format: float
        adaptive_threshold_used:
          type: number
          format: float
        primary_elevated_signal:
          type: string
        pass2_triggered:
          type: boolean
        pass2_trigger_reason:
          type: string
          enum: [pass1_clear, pass1_amber, pass1_high_risk]
        adversarial_elevated_count:
          type: integer
          format: int32
        data_quality:
          type: object
          required:
            - available_signals
            - total_signals
            - availability_ratio
          properties:
            available_signals:
              type: integer
              format: int32
            total_signals:
              type: integer
              format: int32
            availability_ratio:
              type: number
              format: float
    FraudAuditLog:
      type: object
      required:
        - id
        - claim_id
        - worker_id
        - timestamp
        - signals
        - pass1_composite
        - xgboost_score
        - resolution_tier
        - critical_flags
        - model_version
        - adaptive_threshold_applied
      properties:
        id:
          type: string
          format: uuid
        claim_id:
          type: string
          format: uuid
        worker_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        signals:
          type: object
        pass1_composite:
          type: string
        pass2_composite:
          type: string
        xgboost_score:
          type: string
        resolution_tier:
          type: string
        critical_flags:
          type: object
        model_version:
          type: string
        adaptive_threshold_applied:
          type: string